Moved TestingEnvironment to Piwik\Tests\Framework.
