+++ tests/PHPUnit/System/EnvironmentValidationTest.php
@@ -30,7 +30,7 @@ class EnvironmentValidationTest extends SystemTestCase
     {
         parent::setUp();
 
-        $testingEnvironment = new \Piwik_TestingEnvironment();
+        $testingEnvironment = new \Piwik\Tests\Framework\TestingEnvironment();
         $testingEnvironment->configFileGlobal = null;
         $testingEnvironment->configFileLocal = null;
         $testingEnvironment->configFileCommon = null;
@@ -138,7 +138,7 @@ class EnvironmentValidationTest extends SystemTestCase
 
     private function simulateAbsentConfigFile($fileName)
     {
-        $testingEnvironment = new \Piwik_TestingEnvironment();
+        $testingEnvironment = new \Piwik\Tests\Framework\TestingEnvironment();
 
         if ($fileName == 'global.ini.php') {
             $testingEnvironment->configFileGlobal = PIWIK_INCLUDE_PATH . '/tmp/nonexistant/global.ini.php';
@@ -153,7 +153,7 @@ class EnvironmentValidationTest extends SystemTestCase
 
     private function simulateBadConfigFile($fileName)
     {
-        $testingEnvironment = new \Piwik_TestingEnvironment();
+        $testingEnvironment = new \Piwik\Tests\Framework\TestingEnvironment();
 
         if ($fileName == 'global.ini.php') {
             $testingEnvironment->configFileGlobal = PIWIK_INCLUDE_PATH . '/piwik.php';
@@ -168,7 +168,7 @@ class EnvironmentValidationTest extends SystemTestCase
 
     private function simulateHost($host)
     {
-        $testingEnvironment = new \Piwik_TestingEnvironment();
+        $testingEnvironment = new \Piwik\Tests\Framework\TestingEnvironment();
         $testingEnvironment->hostOverride = $host;
         $testingEnvironment->save();
     }
